Use new minicourses and @semapps/core packages
