--- config/ontologies.json
@@ -1,47 +1,0 @@
-[
-  {
-    "prefix": "as",
-    "owl": "",
-    "url": "https://www.w3.org/ns/activitystreams#"
-  },
-  {
-    "prefix": "rdf",
-    "owl": "",
-    "url": "http://www.w3.org/1999/02/22-rdf-syntax-ns#"
-  },
-  {
-    "prefix": "ldp",
-    "owl": "",
-    "url": "http://www.w3.org/ns/ldp#"
-  },
-  {
-    "prefix": "schema",
-    "owl": "",
-    "url": "http://schema.org/"
-  },
-  {
-    "prefix": "sec",
-    "owl": "",
-    "url": "https://w3id.org/security#"
-  },
-  {
-    "prefix": "pair",
-    "owl": "http://virtual-assembly.org/ontologies/pair/ontology.ttl",
-    "url": "http://virtual-assembly.org/ontologies/pair#"
-  },
-  {
-    "prefix": "tutor",
-    "owl": "https://gist.githubusercontent.com/srosset81/be3ea9c820d75995ed967fdac8d8a8ca/raw/ba3ff61d5d3ee6eb6294b6c1b7de385808ecc624/pair-tutor.owl",
-    "url": "http://virtual-assembly.org/ontologies/pair-tutor#"
-  },
-  {
-    "prefix": "semapps",
-    "owl": "",
-    "url": "http://semapps.org/ns/core#"
-  },
-  {
-    "prefix": "oasis",
-    "owl": "",
-    "url": "http://cooperative-oasis.org/ns/core#"
-  }
-]
