+++ moleculer.config.js
@@ -27,5 +27,5 @@ module.exports = {
     CacherMiddleware(cacherConfig), // Set the cacher before the WebAcl middleware
     WebAclMiddleware({ baseUrl: CONFIG.HOME_URL })
   ],
-  // logLevel: 'debug'
+  logLevel: 'debug'
 };
