+++ package.json
@@ -6,25 +6,15 @@
   "scripts": {
     "dev": "moleculer-runner --repl --hot services importers/*.js",
     "start": "moleculer-runner --instances=max services importers/*.js",
-    "link-semapps-packages": "yarn link @semapps/activitypub @semapps/auth @semapps/backup @semapps/fuseki-admin @semapps/importer @semapps/inference @semapps/jsonld @semapps/ldp @semapps/middlewares @semapps/notifications @semapps/mime-types @semapps/signature @semapps/sparql-endpoint @semapps/triplestore @semapps/void @semapps/webacl @semapps/webfinger @semapps/webhooks @semapps/webid"
+    "link-semapps-packages": "yarn link minicourses @semapps/activitypub @semapps/auth @semapps/backup @semapps/core @semapps/importer @semapps/inference @semapps/jsonld @semapps/ldp @semapps/mime-types @semapps/middlewares @semapps/mirror @semapps/notifications @semapps/mime-types @semapps/signature @semapps/sparql-endpoint @semapps/triplestore @semapps/void @semapps/webacl @semapps/webfinger @semapps/webhooks @semapps/webid"
   },
   "dependencies": {
-    "@semapps/activitypub": "0.4.0-alpha.38",
-    "@semapps/auth": "0.4.0-alpha.38",
-    "@semapps/fuseki-admin": "0.4.0-alpha.38",
-    "@semapps/importer": "0.4.0-alpha.38",
-    "@semapps/inference": "0.4.0-alpha.38",
-    "@semapps/jsonld": "0.4.0-alpha.38",
-    "@semapps/ldp": "0.4.0-alpha.38",
-    "@semapps/mime-types": "0.4.0-alpha.38",
-    "@semapps/notifications": "0.4.0-alpha.38",
-    "@semapps/signature": "0.4.0-alpha.38",
-    "@semapps/sparql-endpoint": "0.4.0-alpha.38",
-    "@semapps/triplestore": "0.4.0-alpha.38",
-    "@semapps/void": "0.4.0-alpha.38",
-    "@semapps/webacl": "0.4.0-alpha.38",
-    "@semapps/webfinger": "0.4.0-alpha.38",
-    "@semapps/webid": "0.4.0-alpha.38",
+    "@semapps/auth": "0.4.0-alpha.43",
+    "@semapps/core": "0.4.0-alpha.43",
+    "@semapps/importer": "0.4.0-alpha.43",
+    "@semapps/inference": "0.4.0-alpha.43",
+    "@semapps/notifications": "0.4.0-alpha.43",
+    "@semapps/webid": "0.4.0-alpha.43",
     "@sentry/node": "^5.16.1",
     "@sentry/utils": "^5.16.1",
     "dotenv-flow": "^3.1.0",
@@ -32,6 +22,7 @@
     "handlebars": "^4.7.7",
     "jwt-decode": "^3.1.2",
     "lodash.truncate": "^4.4.2",
+    "minicourses": "1.0.0",
     "moleculer": "^0.14.18",
     "moleculer-bull": "0.2.8",
     "moleculer-cron": "^0.0.2",
