+++ services/inference.service.js
@@ -1,10 +1,10 @@
 const { InferenceService } = require('@semapps/inference');
-const ontologies = require('../config/ontologies.json');
+const { defaultOntologies } = require('@semapps/core');
 
 module.exports = {
   mixins: [InferenceService],
   settings: {
     baseUrl: process.env.SEMAPPS_HOME_URL,
-    ontologies
+    ontologies: defaultOntologies
   }
 };
