+++ services/minicourses.service.js
@@ -1,0 +1,16 @@
+const { MiniCoursesService } = require('minicourses');
+const CONFIG = require('../config/config');
+const transport = require('../config/transport');
+
+module.exports = {
+  mixins: [MiniCoursesService],
+  settings: {
+    baseUrl: CONFIG.HOME_URL,
+    frontendUrl: CONFIG.MINI_COURSES_URL,
+    queueServiceUrl: CONFIG.QUEUE_SERVICE_URL,
+    mailer: {
+      from: `${CONFIG.FROM_NAME} <${CONFIG.FROM_EMAIL}>`,
+      transport
+    }
+  }
+};
