--- services/signature.service.js
@@ -1,9 +1,0 @@
-const path = require('path');
-const { SignatureService } = require('@semapps/signature');
-
-module.exports = {
-  mixins: [SignatureService],
-  settings: {
-    actorsKeyPairsDir: path.resolve(__dirname, '../actors')
-  }
-};
