--- services/sparql-endpoint.service.js
@@ -1,8 +1,0 @@
-const { SparqlEndpointService } = require('@semapps/sparql-endpoint');
-
-module.exports = module.exports = {
-  mixins: [SparqlEndpointService],
-  settings: {
-    ignoreAcl: true
-  }
-};
