--- templates/lesson/subject.hbs
@@ -1,1 +1,0 @@
-{{{ course.pair:label }}}: {{{ lesson.pair:label }}}
