+++ utils.js
@@ -1,5 +1,6 @@
 const createSlug = require('speakingurl');
 const departments = require('./config/departments.json');
+const { ACTOR_TYPES } = require("@semapps/activitypub");
 
 const capitalize = s => (s && s[0].toUpperCase() + s.slice(1)) || "";
 
@@ -45,6 +46,33 @@ const frenchAddressReverseSearch = async (lat, lon) => {
   }
 };
 
+const selectActorData = resource => {
+  let resourceId = resource.id || resource['@id'],
+    resourceTypes = resource.type || resource['@type'];
+  resourceTypes = Array.isArray(resourceTypes) ? resourceTypes : [resourceTypes];
+  if (resourceTypes.includes('foaf:Person')) {
+    return {
+      '@type': ACTOR_TYPES.PERSON,
+      name: (resource['foaf:name'] && resource['foaf:familyName']) ? resource['foaf:name'] + ' ' + resource['foaf:familyName'] : undefined,
+      preferredUsername: getSlugFromUri(resourceId)
+    };
+  } else if (resourceTypes.includes('pair:Organization')) {
+    return {
+      '@type': ACTOR_TYPES.ORGANIZATION,
+      name: resource['pair:label'],
+      preferredUsername: getSlugFromUri(resourceId)
+    };
+  } else if (resourceTypes.includes('pair:Group')) {
+    return {
+      type: ACTOR_TYPES.GROUP,
+      name: resource['pair:label'],
+      preferredUsername: getSlugFromUri(resourceId)
+    };
+  } else {
+    return false;
+  }
+};
+
 module.exports = {
   capitalize,
   getSlugFromUri,
@@ -53,5 +81,6 @@ module.exports = {
   distanceBetweenPoints,
   addDays,
   removeTime,
-  frenchAddressReverseSearch
+  frenchAddressReverseSearch,
+  selectActorData
 };
